refactor(tests): consolidate 9 test files into 4 by module, fix stale MNN assertions
