--- tests/README.md
@@ -1,3 +1,0 @@
-# Tests
-
-Repository-level test notes and validation plans.
