+++ tests/test_e2e.sh
@@ -4,4 +4,4 @@ set -euo pipefail
 
 SCRIPT_DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
 echo "[omniinfer-flow] shell host: $(uname -s) $(uname -m)"
-exec python3 "$SCRIPT_DIR/test_inference_flow.py" "$@"
+exec python3 "$SCRIPT_DIR/test_e2e.py" "$@"
